Integrate GUI with the application.
Issue #15: I have added a JComboBox to select the type of generator which should process the input text, along with a separate text area to display the result. I have also added an event listener to the 'Generate Text' button. When the button is clicked, the application will create an object of the selected generator type with the text in the input area as a parameter to its constructor, and call the generateText() method. I have selected a word limit of 200 here, but I suppose a text field to enter the word limit could be added to the GUI at a later stage. Finally, I have added a JLabel at the bottom to display the time taken for execution (in milliseconds). The label contents are updated each time the 'Generate Text' button is clicked.